[BE] refactor : openApi Url 분리
- OpenApiUrl.java : 클래스 분리했습니다.

associate with #17
