+++ BE/dust12/src/main/java/dev/codesquad/java/dust12/ApiParam.java
@@ -1,9 +1,5 @@
 package dev.codesquad.java.dust12;
 
-import java.time.ZoneId;
-import java.time.ZonedDateTime;
-import java.time.format.DateTimeFormatter;
-
 public class ApiParam {
     // KAKAO COORDINATE API
     public static final String JSON_DOCUMENTS = "documents";
